+++ geohash/geohash.h
@@ -15,16 +15,16 @@
 bool GEOHASH_verify(const std::string & hash);
 
 // encode a point, return hash.
-std::string GEOHASH_encode(const dPoint & p, unsigned int hash_length);
+std::string GEOHASH_encode(const dPoint & p, size_t hash_length);
 
 // Encode a rectangle, return longest hash (but not longer then maxlen)
 // which covers the rectangle. Can return an ampty string.
-std::string GEOHASH_encode(const dRect & r, unsigned int maxlen);
+std::string GEOHASH_encode(const dRect & r, size_t maxlen);
 
 // Encode a rectangle, return up to 4 longest adjacent hashes (not
 // longer then maxlen and non-empty) which covers the rectangle. Should be
 // efficient (give longer hashas) for rectangles which cross hash bundaries.
-std::set<std::string> GEOHASH_encode4(const dRect & r, unsigned int maxlen);
+std::set<std::string> GEOHASH_encode4(const dRect & r, size_t maxlen);
 
 // Decode a hash, return coordinate range. Range is empty if hash is not valid.
 dRect GEOHASH_decode(const std::string & hash);
